+++ frontend/settings.lua
@@ -70,18 +70,6 @@ function dump(data)
 	return table.concat(out)
 end
 
-function DEBUG(...)
-	local line = ""
-	for i,v in ipairs({...}) do
-		if type(v) == "table" then
-			line = line .. " " .. dump(v)
-		else
-			line = line .. " " .. tostring(v)
-		end
-	end
-	print("#"..line)
-end
-
 -- simple serialization function, won't do uservalues, functions, loops
 function DocSettings:_serialize(what, outt, indent)
 	if type(what) == "table" then
