+++ frontend/ui/reader/readerbookmark.lua
@@ -1,4 +1,4 @@
-require "ui/notification"
+require "ui/widget/notification"
 
 ReaderBookmark = InputContainer:new{
 	bm_menu_title = "Bookmarks",
@@ -89,7 +89,7 @@ function ReaderBookmark:onShowBookmark()
 	local bm_menu = Menu:new{
 		title = "Bookmarks",
 		item_table = self.bookmarks,
-		width = Screen:getWidth()-20, 
+		width = Screen:getWidth()-20,
 		height = Screen:getHeight(),
 	}
 	-- buid up menu widget method as closure
@@ -111,7 +111,7 @@ function ReaderBookmark:onShowBookmark()
 		dimen = Screen:getSize(),
 		bm_menu,
 	}
-	bm_menu.close_callback = function() 
+	bm_menu.close_callback = function()
 		UIManager:close(menu_container)
 	end
 
@@ -154,7 +154,7 @@ function ReaderBookmark:addBookmark(pn_or_xp)
 		return self:isBookmarkInSequence(a, b)
 	end)
 	return true
-end 
+end
 
 function ReaderBookmark:isBookmarkInSequence(a, b)
 	return a.page < b.page
