+++ frontend/ui/reader/readerdogear.lua
@@ -1,3 +1,4 @@
+require "ui/widget/image"
 
 ReaderDogear = RightContainer:new{}
