+++ frontend/ui/reader/readerfooter.lua
@@ -1,3 +1,4 @@
+require "ui/widget/progress"
 
 ReaderFooter = InputContainer:new{
 	pageno = nil,
