+++ frontend/ui/widget/config.lua
@@ -1,26 +1,5 @@
-require "ui/widget"
-require "ui/focusmanager"
-require "ui/infomessage"
-require "ui/font"
-require "ui/toggleswitch"
-
-FixedTextWidget = TextWidget:new{}
-function FixedTextWidget:getSize()
-	local tsize = sizeUtf8Text(0, Screen:getWidth(), self.face, self.text, true)
-	if not tsize then
-		return Geom:new{}
-	end
-	self._length = tsize.x
-	self._height = self.face.size
-	return Geom:new{
-		w = self._length,
-		h = self._height,
-	}
-end
-
-function FixedTextWidget:paintTo(bb, x, y)
-	renderUtf8Text(bb, x, y+self._height, self.face, self.text, true)
-end
+require "ui/widget/container"
+require "ui/widget/toggleswitch"
 
 MenuBarItem = InputContainer:new{}
 function MenuBarItem:init()
@@ -190,7 +169,7 @@ function ConfigOption:init()
 			local item_font_face = self.options[c].item_font_face and self.options[c].item_font_face or "cfont"
 			local item_font_size = self.options[c].item_font_size and self.options[c].item_font_size or default_item_font_size
 			local option_height = (self.options[c].height and self.options[c].height or default_option_height) * Screen:getDPI()/167
-			local items_spacing = HorizontalSpan:new{ 
+			local items_spacing = HorizontalSpan:new{
 				width = (self.options[c].spacing and self.options[c].spacing or default_items_spacing) * Screen:getDPI()/167
 			}
 			local horizontal_group = HorizontalGroup:new{}
@@ -205,7 +184,7 @@ function ConfigOption:init()
 				table.insert(option_name_container, option_name)
 				table.insert(horizontal_group, option_name_container)
 			end
-			
+
 			if self.options[c].widget == "ProgressWidget" then
 				local widget_container = CenterContainer:new{
 					dimen = Geom:new{w = Screen:getWidth()*self.options[c].widget_align_center, h = option_height}
@@ -218,7 +197,7 @@ function ConfigOption:init()
 				table.insert(widget_container, widget)
 				table.insert(horizontal_group, widget_container)
 			end
-			
+
 			local option_items_container = CenterContainer:new{
 				dimen = Geom:new{w = Screen:getWidth()*item_align, h = option_height}
 			}
@@ -261,7 +240,7 @@ function ConfigOption:init()
 					end
 				end
 			end
-			
+
 			if self.options[c].item_text then
 				for d = 1, #self.options[c].item_text do
 					local option_item = nil
@@ -298,7 +277,7 @@ function ConfigOption:init()
 					end
 				end
 			end
-			
+
 			if self.options[c].item_icons then
 				for d = 1, #self.options[c].item_icons do
 					local option_item = OptionIconItem:new{
@@ -322,7 +301,7 @@ function ConfigOption:init()
 					end
 				end
 			end
-			
+
 			if self.options[c].toggle then
 				local switch = ToggleSwitch:new{
 					name = self.options[c].name,
@@ -338,7 +317,7 @@ function ConfigOption:init()
 				switch:setPosition(position)
 				table.insert(option_items_group, switch)
 			end
-			
+
 			table.insert(option_items_container, option_items_group)
 			table.insert(horizontal_group, option_items_container)
 			table.insert(vertical_group, horizontal_group)
@@ -352,7 +331,7 @@ end
 ConfigPanel = FrameContainer:new{ background = 0, bordersize = 0, }
 function ConfigPanel:init()
 	local config_options = self.config_dialog.config_options
-	local default_option = config_options.default_options and config_options.default_options 
+	local default_option = config_options.default_options and config_options.default_options
 							or config_options[1].options
 	local panel = ConfigOption:new{
 		options = self.index and config_options[self.index].options or default_option,
@@ -375,26 +354,26 @@ function MenuBar:init()
 		local icon_dimen = menu_icon:getSize()
 		icons_width = icons_width + icon_dimen.w
 		icons_height = icon_dimen.h > icons_height and icon_dimen.h or icons_height
-		
+
 		menu_items[c] = MenuBarItem:new{
 			menu_icon,
 			index = c,
 			config = self.config_dialog,
 		}
 	end
-	
+
 	local spacing = HorizontalSpan:new{
 		width = (Screen:getWidth() - icons_width) / (#menu_items+1)
 	}
-	
+
 	local menu_bar = HorizontalGroup:new{}
-	
+
 	for c = 1, #menu_items do
 		table.insert(menu_bar, spacing)
 		table.insert(menu_bar, menu_items[c])
 	end
 	table.insert(menu_bar, spacing)
-	
+
 	self.dimen = Geom:new{ w = Screen:getWidth(), h = icons_height}
 	table.insert(self, menu_bar)
 end
@@ -415,7 +394,7 @@ Widget that displays config menubar and config panel
  +----------------+
  |    Menu Bar    |
  +----------------+
- 
+
 --]]
 
 ConfigDialog = InputContainer:new{
@@ -429,7 +408,7 @@ function ConfigDialog:init()
 	self.config_panel = ConfigPanel:new{
 		config_dialog = self,
 	}
-	self.config_menubar = MenuBar:new{ 
+	self.config_menubar = MenuBar:new{
 		config_dialog = self,
 	}
 	self:makeDialog()
@@ -454,7 +433,7 @@ function ConfigDialog:init()
 		self.key_events.FocusRight = nil
 	end
 	self.key_events.Select = { {"Press"}, doc = "select current menu item"}
-	
+
 	UIManager:setDirty(self, "partial")
 end
 
@@ -470,9 +449,9 @@ function ConfigDialog:makeDialog()
 		self.config_panel,
 		self.config_menubar,
 	}
-	
+
 	local dialog_size = dialog:getSize()
-	
+
 	self[1] = BottomContainer:new{
 		dimen = Screen:getSize(),
 		FrameContainer:new{
@@ -481,13 +460,13 @@ function ConfigDialog:makeDialog()
 			dialog,
 		}
 	}
-	
+
 	self.dialog_dimen = Geom:new{
 		x = (Screen:getWidth() - dialog_size.w)/2,
 		y = Screen:getHeight() - dialog_size.h,
 		w = dialog_size.w,
 		h = dialog_size.h,
-	}	
+	}
 end
 
 function ConfigDialog:onShowConfigPanel(index)
